implementação de servicos

+++ src/org/iftm/poo/boundary/LivroDTO.java
@@ -1,6 +1,7 @@
 package org.iftm.poo.boundary;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import org.iftm.poo.model.domain.Livro;
 
@@ -21,11 +22,11 @@ public class LivroDTO implements Serializable {
 	private String nomeCategoria;
 	
 	private String nomeImagem;
-	private Double altura;
-	private Double largura;
-	private Double comprimento;
-	private Double diametro;
-	private Double peso;
+	private BigDecimal altura;
+	private BigDecimal largura;
+	private BigDecimal comprimento;
+	private BigDecimal diametro;
+	private BigDecimal peso;
 	
 	public LivroDTO(){
 		
@@ -148,43 +149,43 @@ public class LivroDTO implements Serializable {
 		this.nomeImagem = nomeImagem;
 	}
 
-	public Double getAltura() {
+	public BigDecimal getAltura() {
 		return altura;
 	}
 
-	public void setAltura(Double altura) {
+	public void setAltura(BigDecimal altura) {
 		this.altura = altura;
 	}
 
-	public Double getLargura() {
+	public BigDecimal getLargura() {
 		return largura;
 	}
 
-	public void setLargura(Double largura) {
+	public void setLargura(BigDecimal largura) {
 		this.largura = largura;
 	}
 
-	public Double getComprimento() {
+	public BigDecimal getComprimento() {
 		return comprimento;
 	}
 
-	public void setComprimento(Double comprimento) {
+	public void setComprimento(BigDecimal comprimento) {
 		this.comprimento = comprimento;
 	}
 
-	public Double getDiametro() {
+	public BigDecimal getDiametro() {
 		return diametro;
 	}
 
-	public void setDiametro(Double diametro) {
+	public void setDiametro(BigDecimal diametro) {
 		this.diametro = diametro;
 	}
 
-	public Double getPeso() {
+	public BigDecimal getPeso() {
 		return peso;
 	}
 
-	public void setPeso(Double peso) {
+	public void setPeso(BigDecimal peso) {
 		this.peso = peso;
 	}
